requirements.txt
flask
geopandas
rasterio
folium
pystac-client
planetary-computer
shapely
Pillow 
numpy
matplotlib
ultralytics